Extract composer runtime helpers
